meeting-sms/requirements.txt
flask==3.1.*
gunicorn==23.*
gspread==6.*
google-auth==2.*
twilio==9.*
